% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_map.R
\name{as.list.neer_match.similarity_map.SimilarityMap}
\alias{as.list.neer_match.similarity_map.SimilarityMap}
\title{Convert a similarity map to a list}
\usage{
\method{as.list}{neer_match.similarity_map.SimilarityMap}(x, ...)
}
\arguments{
\item{x}{A similarity map object.}

\item{...}{Additional arguments (not used).}
}
\value{
A list of instructions.
}
\description{
Returns a nested parsed list of instructions that define how to compare two
data sets. Each instruction (outer list element) has three parts (inner
list). The first part is the left field name, the second part is the right
field, and the third part is a similarity function.
}
\examples{
smap <- SimilarityMap(list(
  "title" = list("jaro"),
  "author~who" = list("lcsseq", "hamming")
))
as.list(smap)
}
