% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_map.R
\name{association_names}
\alias{association_names}
\alias{association_names,neer_match.similarity_map.SimilarityMap-method}
\title{Association names}
\usage{
association_names(object)

\S4method{association_names}{neer_match.similarity_map.SimilarityMap}(object)
}
\arguments{
\item{object}{A similarity map object.}
}
\value{
A list of association names.
}
\description{
A list with field pair names separated by underscores.
}
\examples{
smap <- SimilarityMap(list(
  "title" = list("jaro"),
  "author~who" = list("lcsseq", "hamming")
))
association_names(smap)
}
