% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_map.R
\name{association_offsets}
\alias{association_offsets}
\alias{association_offsets,neer_match.similarity_map.SimilarityMap-method}
\title{Association offsets}
\usage{
association_offsets(object)

\S4method{association_offsets}{neer_match.similarity_map.SimilarityMap}(object)
}
\arguments{
\item{object}{A similarity map object.}
}
\value{
A list of offsets.
}
\description{
A list with similarity matrix offsets for each field pair.
}
\examples{
smap <- SimilarityMap(list(
  "title" = list("jaro"),
  "author~who" = list("lcsseq", "hamming")
))
association_offsets(smap)
}
