% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_map.R
\name{association_sizes}
\alias{association_sizes}
\alias{association_sizes,neer_match.similarity_map.SimilarityMap-method}
\title{Association sizes}
\usage{
association_sizes(object)

\S4method{association_sizes}{neer_match.similarity_map.SimilarityMap}(object)
}
\arguments{
\item{object}{A similarity map object.}
}
\value{
A list of sizes.
}
\description{
A list with the number of similarity functions for each field pair.
}
\examples{
smap <- SimilarityMap(list(
  "title" = list("jaro"),
  "author~who" = list("lcsseq", "hamming")
))
association_sizes(smap)
}
