% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_map.R
\name{available_similarities}
\alias{available_similarities}
\title{Predefined available similarities}
\usage{
available_similarities()
}
\value{
A list of predefined available similarities.
}
\description{
Predefined available similarities
}
\examples{
available_similarities()
}
