% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_model.R
\name{DLMatchingModel}
\alias{DLMatchingModel}
\title{Deep Learning Matching Model}
\usage{
DLMatchingModel(
  similarity_map,
  initial_feature_width_scales = 10L,
  feature_depths = 2L,
  initial_record_width_scale = 10L,
  record_depth = 4L,
  ...
)
}
\arguments{
\item{similarity_map}{A \code{\link{SimilarityMap}} object.}

\item{initial_feature_width_scales}{An integer or an integer vector of
initial feature width scales for each field-pair network. The scale is
multiplied by the number of similarities used in the field-pair network
to determine the number of units of the first dense layer. If the input is a
scalar, the same value is used for all field-pair networks.}

\item{feature_depths}{An integer or an integer vector of feature depths for
each field-pair network. The depth is the number of hidden dense layers
used in the field-pair network. If the input is a scalar, the same value is
used for all field-pair networks.}

\item{initial_record_width_scale}{An integer representing the initial record
width scale. The scale is multiplied by the number of field-pair networks
to determine the number of units of the first dense layer of the record-pair
network.}

\item{record_depth}{An integer representing the record depth. The depth is
the number of hidden dense layers used in the record-pair network.}

\item{...}{Additional arguments passed to the Python constructor. These
arguments are passed down to the
\href{https://www.tensorflow.org/api_docs/python/tf/keras/Model}{
\code{tf.keras.Model}} constructor.}
}
\description{
Construct a deep learning entity-matching model based on field similarity
encodings. The class wraps the constructor of the \code{DLMatchingModel} Python
class from the \code{neer-match} package. The class is built using the
\code{\link{tensorflow}} framework.
}
\examples{
smap <- SimilarityMap(
  instructions = list(
    `a~c` = list("discrete", "gaussian"),
    `b` = list("discrete", "levenshtein")
  )
)
model <- DLMatchingModel(smap)
}
\seealso{
\code{\link{SimilarityMap}}, \code{\link{NSMatchingModel}}
}
