% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_model.R
\name{fit}
\alias{fit}
\alias{fit,neer_match.matching_model.DLMatchingModel-method}
\alias{fit,neer_match.matching_model.NSMatchingModel-method}
\title{Model Training}
\usage{
fit(object, left, right, matches, ...)

\S4method{fit}{neer_match.matching_model.DLMatchingModel}(object, left, right, matches, ...)

\S4method{fit}{neer_match.matching_model.NSMatchingModel}(
  object,
  left,
  right,
  matches,
  epochs,
  satisfiability_weight = 1,
  verbose = 1L,
  log_mod_n = 1L,
  ...
)
}
\arguments{
\item{object}{A matching model object.}

\item{left}{A data frame with the left records.}

\item{right}{A data frame with the right records.}

\item{matches}{A data frame with the indices of the matching record pairs.}

\item{...}{Additional arguments passed to
\href{https://www.tensorflow.org/api_docs/python/tf/keras/Model#fit}{
tf.keras.fit}.}

\item{epochs}{The number of epochs to train the model.}

\item{satisfiability_weight}{A numeric value in the range \eqn{[0, 1]}
representing the weight allocated to the satisfiability loss of a
hybrid model.}

\item{verbose}{An integer indicating the verbosity level.}

\item{log_mod_n}{An positive integer that determines the frequency of
logging. The method logs every \code{log_mod_n} epochs.}
}
\value{
Called for side effects (model training).
}
\description{
Fits the model. The method emulates the behavior of the
\href{https://www.tensorflow.org/api_docs/python/tf/keras/Model#fit}{
tf.keras.fit} method. It automatically constructs a data generator
from the left and right datasets iterating over all the elements of their
Cartesian product. The generator's labels are generated from the matches
data frame. The method uses the generator to train the model.

\subsection{\code{\link{DLMatchingModel}}}{
The method passes the constructed generator and any
additional call arguments to
directly to the
\href{https://www.tensorflow.org/api_docs/python/tf/keras/Model#fit}{
tf.keras.fit}.
}

\subsection{\code{\link{DLMatchingModel}}}{
The method passes the constructed generator and any
additional call arguments to
directly to the
\href{https://www.tensorflow.org/api_docs/python/tf/keras/Model#fit}{
tf.keras.fit}.
}
}
\examples{
smap <- SimilarityMap(
  instructions = list(
    `score` = list("gaussian", "euclidean"),
    `platform` = list("osa", "indel")
  )
)
model <- NSMatchingModel(smap)
compile(model)
matching_data <- fuzzy_games_example_data()
fit(
  model,
  matching_data$left, matching_data$right, matching_data$matches,
  epochs = 1L
)
}
