% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{fuzzy_games_example_data}
\alias{fuzzy_games_example_data}
\title{Fuzzy Games Matching Example Data}
\usage{
fuzzy_games_example_data()
}
\value{
A left data frame with 36 rows and 6 columns. A right data frame
with 39 rows and 6 columns. A matches data frame with 39 rows and 2 columns.
}
\description{
Returns three datasets used in the
\code{vignette("fuzzy_games", package = "neermatch")} vignette.

The left dataset is a subset of the \code{game_reviews} data containing games
with titles starting with "Metal Gear" or "Metal Slug". The selection
results in 36 records, many of which have the exact same \code{title},
\code{developer}, and release \code{year} and differ only in their release
\code{platform}, \code{score}, and \code{reviews} fields.

The initial selection of the right records is the same as in the left,
but the data of the right dataset introduce noise in the \code{title} and
\code{developer} columns. Three characters are randomly removed from the
\code{title} column, and one character is randomly removed from the
\code{developer} column. The mutated \code{developer} column is stored in the
column \code{dev} of the right dataset. In addition, three records of the
left dataset are randomly duplicated in the right dataset.

The matches dataset is a data frame with the matching indices of the
left and right datasets.
}
\examples{
matching_data <- fuzzy_games_example_data()
head(matching_data$left)
head(matching_data$right)
head(matching_data$matches)
}
\seealso{
\code{\link{game_reviews}}
}
\keyword{datasets}
\keyword{games}
\keyword{reviews}
