% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{game_reviews}
\alias{game_reviews}
\title{Game Reviews Data}
\format{
A data frame with 47201 rows and 6 columns
}
\source{
\href{https://docs.google.com/spreadsheets/d/1wtV8yBr5RXAjO_1kakzFcHd7TKGw7pIUVNI5SYouCUM}{
GameRankings.com data dump on 2019-12-08}
}
\usage{
data(game_reviews)
}
\description{
Game Reviews Data
}
\details{
\subsection{Data Description.}{
The data is a copy of an archived version of the game ranking data of
GameRankings.com before it shut down. The snapshot was taken 2019-12-08.
The variables are as follows:

\itemize{
\item \code{title (character)} The game title.
\item \code{platform (character)} The release platform of the game.
\item \code{year (integer)} The year the game was released.
\item \code{score (numeric)} The average user review score of the game.
\item \code{reviews (integer)} The number of user reviews.
\item \code{developer (character)} The developer of the game.
}
}
}
\examples{
data(game_reviews)
head(game_reviews)
}
\keyword{datasets}
\keyword{games}
\keyword{reviews}
