% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_map.R
\name{no_associations}
\alias{no_associations}
\alias{no_associations,neer_match.similarity_map.SimilarityMap-method}
\title{Number of associations}
\usage{
no_associations(object)

\S4method{no_associations}{neer_match.similarity_map.SimilarityMap}(object)
}
\arguments{
\item{object}{A similarity map object.}
}
\value{
The number of associations.
}
\description{
The number of associations in the similarity map.
}
\examples{
smap <- SimilarityMap(list(
  "title" = list("jaro"),
  "author~who" = list("lcsseq", "hamming")
))
no_associations(smap)
}
