% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reasoning.R
\name{RefutationModel}
\alias{RefutationModel}
\title{Refutation Model}
\usage{
RefutationModel(
  similarity_map,
  initial_feature_width_scales = 10L,
  feature_depths = 2L,
  initial_record_width_scale = 10L,
  record_depth = 4L,
  ...
)
}
\arguments{
\item{similarity_map}{A \code{\link{SimilarityMap}} object.}

\item{initial_feature_width_scales}{An integer or an integer vector of
initial feature width scales for each field-pair network. The scale is
multiplied by the number of similarities used in the field-pair network
to determine the number of units of the first dense layer. If the input is a
scalar, the same value is used for all field-pair networks.}

\item{feature_depths}{An integer or an integer vector of feature depths for
each field-pair network. The depth is the number of hidden dense layers
used in the field-pair network. If the input is a scalar, the same value is
used for all field-pair networks.}

\item{initial_record_width_scale}{An integer representing the initial record
width scale. The scale is multiplied by the number of field-pair networks
to determine the number of units of the first dense layer of the record-pair
network.}

\item{record_depth}{An integer representing the record depth. The depth is
the number of hidden dense layers used in the record-pair network.}

\item{...}{Additional arguments passed to the Python constructor. These
arguments are passed down to the
\href{https://www.tensorflow.org/api_docs/python/tf/keras/Model}{
\code{tf.keras.Model}} constructor.
constructor.}
}
\description{
todo
}
\examples{
smap <- SimilarityMap(
  instructions = list(
    `z~w` = list("jaro", "levenshtein", "discrete"),
    `b ~ c` = list("jaro_winkler", "hamming")
  )
)
model <- RefutationModel(smap)
}
\seealso{
\code{\link{SimilarityMap}}
}
