% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_map.R
\name{show,neer_match.similarity_map.SimilarityMap-method}
\alias{show,neer_match.similarity_map.SimilarityMap-method}
\title{Print the similarity map to standard output}
\usage{
\S4method{show}{neer_match.similarity_map.SimilarityMap}(object)
}
\arguments{
\item{object}{A similarity map object.}
}
\value{
Called for its side effect (print to standard output).
}
\description{
Print the similarity map to standard output
}
\examples{
smap <- SimilarityMap(list(
  "title" = list("damerau_levenshtein", "jaro"),
  "author~who" = list("lcsseq", "hamming"),
  "when~date" = list("euclidean")
))
show(smap)
}
