% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_map.R
\name{similarity_keys}
\alias{similarity_keys}
\alias{similarity_keys,neer_match.similarity_map.SimilarityMap-method}
\title{Similarity keys}
\usage{
similarity_keys(object)

\S4method{similarity_keys}{neer_match.similarity_map.SimilarityMap}(object)
}
\arguments{
\item{object}{A similarity map object.}
}
\value{
A list of keys.
}
\description{
A list combining association names and similarity functions seperated by
underscores.
}
\examples{
smap <- SimilarityMap(list(
  "title" = list("jaro"),
  "author~who" = list("lcsseq", "hamming")
))
similarity_keys(smap)
}
